Change re. Romans 13:12-14
"The night is nearly over; the day has drawn near. So let us lay aside
the deeds of darkness and put on the armor of light. Let us behave
decently, as in the daytime, not in carousing and drunkenness, not
in sexual immorality and debauchery, not in dissension and jealousy.
Instead, clothe yourselves with the Lord Jesus Christ, and make no
provision for the desires of the flesh."